Move composite experiments folder to library
