+++ docs/apidocs/composite.rst
@@ -1,6 +1,6 @@
 .. _qiskit-experiments-composite:
 
-.. automodule:: qiskit_experiments.composite
+.. automodule:: qiskit_experiments.library.composite
    :no-members:
    :no-inherited-members:
    :no-special-members:
