+++ qiskit_experiments/library/randomized_benchmarking/rb_experiment.py
@@ -25,7 +25,7 @@ from qiskit.circuit import Gate
 
 from qiskit_experiments.base_experiment import BaseExperiment
 from qiskit_experiments.analysis.data_processing import probability
-from qiskit_experiments.composite import ParallelExperiment
+from qiskit_experiments.library.composite import ParallelExperiment
 from .rb_analysis import RBAnalysis
 from .clifford_utils import CliffordUtils
 from .rb_utils import RBUtils
