+++ qiskit_experiments/test/t2ramsey_backend.py
@@ -14,8 +14,6 @@ T2Ramsey Backend class.
 Temporary backend to be used for t2ramsey experiment
 """
 
-# pylint: disable = unused-import
-
 import numpy as np
 
 from qiskit.utils import apply_prefix
@@ -23,9 +21,6 @@ from qiskit.providers import BackendV1
 from qiskit.providers.options import Options
 from qiskit.providers.models import QasmBackendConfiguration
 from qiskit.result import Result
-from qiskit.test import QiskitTestCase
-from qiskit_experiments.composite import ParallelExperiment
-from qiskit_experiments.library import T2Ramsey
 from qiskit_experiments.test.mock_job import MockJob
 
 # Fix seed for simulations
