+++ test/calibration/experiments/test_rabi.py
@@ -29,7 +29,7 @@ from qiskit_experiments.library import Rabi, EFRabi
 from qiskit_experiments.library.calibration.analysis.oscillation_analysis import OscillationAnalysis
 from qiskit_experiments.data_processing.data_processor import DataProcessor
 from qiskit_experiments.data_processing.nodes import Probability
-from qiskit_experiments.composite.parallel_experiment import ParallelExperiment
+from qiskit_experiments.library import ParallelExperiment
 from qiskit_experiments.test.mock_iq_backend import MockIQBackend
